fix: two pre-existing bugs surfaced by this branch's first real CI run
Neither is related to the Android work - this branch had simply never
had a real GitHub Actions run before this PR, so both slipped through
local (Windows/MSVC) testing:

- mdct.cpp: ForwardCosTable's NLen/kHalf (int) implicitly convert to
  std::array's std::size_t template argument, which GCC's
  -Wsign-conversion (but not MSVC) flags under -Werror. Explicit
  static_cast at both use sites.

- assignment.hpp: Assignment/route()/dual_mono_routing()/derive_codec()/
  format_destination()/parse_destination() were missing AC3FORGE_EXPORT
  entirely (unlike every other public symbol in plan.hpp beside it),
  invisible with forge_objects's hidden-by-default visibility - only
  caught by the new BUILD_SHARED_LIBS=ON verification pass linking
  ac3tests against the real .so. Verified locally against a real
  Windows shared-lib build (dllexport model, same class of bug):
  297/298 tests pass (the one failure is the same pre-existing
  environment-dependent render-device-enumeration test).

+++ src/lib/src/core/mdct.cpp
@@ -72,7 +72,8 @@ const Twiddles2& twiddles2() {
 template <int NLen>
 struct ForwardCosTable {
     static constexpr int kHalf = NLen / 2;
-    std::array<std::array<double, NLen>, kHalf> value{};
+    std::array<std::array<double, static_cast<std::size_t>(NLen)>, static_cast<std::size_t>(kHalf)>
+        value{};
     explicit ForwardCosTable(double alpha) {
         for (int k = 0; k < kHalf; ++k) {
             const double factor = 2.0 * k + 1.0;
